avro-people-service/requirements.txt
confluent-kafka==2.3.0
fastavro==1.9.4
fastapi[all]==0.111.0
python-dotenv==1.0.1
Faker==25.0.1
requests==2.31.0